Updated logic of main method

+++ weather-driven-booking-engine/src/main/java/businessmvc/model/ResponseToUserPetition.java
@@ -1,6 +1,6 @@
 package businessmvc.model;
 
 public record ResponseToUserPetition(String hotel, String island, String checkIn, String checkOut,
-                                     Float totalRate, Float meanTemp, Float meanRain, Integer meanHumidity,
-                                     Integer meanClouds, Float meanWindSpeed) {
+                                     Float totalRate, Float meanTemp, Float meanRain, Float meanHumidity,
+                                     Float meanClouds, Float meanWindSpeed) {
 }
